requirement.txt
envbash==1.2.0

flask==1.1.1
Flask-RESTful==0.3.7
flask-cors==3.0.8
Flask-JWT-Extended==3.24.1
gunicorn==20.0.4

marshmallow==3.5.1
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
PyMySQL==0.9.3
marshmallow-sqlalchemy==0.22.3